@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -8,17 +8,17 @@ class Pizza extends Model
 {
     public function set()
     {
-        return $this->belongsTo('App\PizzaSet');
+        return $this->belongsTo('App\Models\PizzaSet');
     }
 
 
     public function order()
     {
-        return $this->belongsTo('App\Order');
+        return $this->belongsTo('App\Models\Order');
     }
 
     public function comments()
     {
-        return $this->morphToMany('App\Comment', 'commentable');
+        return $this->morphToMany('App\Models\Comment', 'commentable');
     }
 }
