@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -8,6 +8,6 @@ class Role extends Model
 {
     public function employees()
     {
-        return $this->hasMany('App\Employee');
+        return $this->hasMany('App\Models\Employee');
     }
 }
