@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -8,12 +8,12 @@ class Stock extends Model
 {
     public function products()
     {
-        return $this->belongsToMany('App\Product');
+        return $this->belongsToMany('App\Models\Product');
     }
 
 
     public function comments()
     {
-        return $this->morphToMany('App\Comment', 'commentable');
+        return $this->morphToMany('App\Models\Comment', 'commentable');
     }
 }
