@@ -25,14 +25,16 @@
                                     <div class="row">
 
                                         <div class="col-md-12">
-                                            <img v-if="!prodImageChanged" class="prod-details-image" 
-                                                :src="prodData.image_thumbs.w_600" alt="prod image">
-                                            <input v-if="prodImageChanged" type="file" id="productImage" ref="prodImageFile"
-                                                @change="handleFileUpload">
+                                            <img v-if="!prodImageChanged" class="prod-details-image"
+                                                 :src="prodData.image_thumbs.w_600" alt="prod image">
+                                            <input v-if="prodImageChanged" type="file" id="productImage"
+                                                   ref="prodImageFile"
+                                                   @change="handleFileUpload">
                                         </div>
 
                                         <div class="col-md-12 mt-5">
-                                            <button v-if="!prodImageChanged" class="btn btn-default" @click="changeImage">
+                                            <button v-if="!prodImageChanged" class="btn btn-default"
+                                                    @click="changeImage">
                                                 <i class="fa fa-edit"></i> Change
                                             </button>
                                         </div>
@@ -41,7 +43,7 @@
                                 </div>
 
                                 <div class="col-md-12 mt-15">
-                                    
+
                                     <div class="col-12">
                                         <div class="form-group">
                                             <label>Name</label>
@@ -56,10 +58,10 @@
 
                                                 <div class="form-group">
                                                     <label for="prodTypesSelect">Type</label>
-                                                    <select v-model="prodEdit.type_id" class="form-control" 
-                                                        id="prodTypesSelect">
+                                                    <select v-model="prodEdit.type_id" class="form-control"
+                                                            id="prodTypesSelect">
                                                         <option v-for="type in prodTypesList" :key="type.id"
-                                                            :value="type.id">
+                                                                :value="type.id">
                                                             {{ type.name }}
                                                         </option>
                                                     </select>
@@ -77,8 +79,8 @@
 
                                                 <div class="form-group">
                                                     <label>Cost ($)</label>
-                                                    <input type="number" step="0.01" min="0.01" 
-                                                        class="form-control" v-model="prodEdit.cost">
+                                                    <input type="number" step="0.01" min="0.01"
+                                                           class="form-control" v-model="prodEdit.cost">
                                                 </div>
 
                                             </div>
@@ -86,8 +88,8 @@
 
                                                 <div class="form-group">
                                                     <label>Weight (g)</label>
-                                                    <input type="number" step="1" min="1" 
-                                                        class="form-control" v-model="prodEdit.weight">
+                                                    <input type="number" step="1" min="1"
+                                                           class="form-control" v-model="prodEdit.weight">
                                                 </div>
 
                                             </div>
@@ -99,7 +101,8 @@
 
                                         <div class="form-group">
                                             <label>Description</label>
-                                            <textarea rows="5" class="form-control" v-model="prodEdit.description"></textarea>
+                                            <textarea rows="5" class="form-control"
+                                                      v-model="prodEdit.description"></textarea>
                                         </div>
 
                                     </div>
@@ -110,11 +113,11 @@
                                     <div class="col-md-12 text-right">
 
                                         <button class="btn btn-success btn-fill btn-icon"
-                                            @click="saveProd"> 
+                                                @click="saveProd">
                                             <i class="fa fa-check"></i> Save
                                         </button>
                                         <button class="btn btn-warning btn-fill btn-icon"
-                                            @click="cancelEdit"> 
+                                                @click="cancelEdit">
                                             <i class="fa fa-ban"></i> Cancel
                                         </button>
 
@@ -196,7 +199,7 @@
                     formData.append(prop, this.prodEdit[prop]);
                 }
 
-                axios.post('/products/save-product',
+                axios.post('/products/save',
                     formData,
                     {
                         headers: {
@@ -204,16 +207,16 @@
                         }
                     }
                 ).then(function(response) {
-                    
+
                     this.notifySuccess('Product ID:' + this.prodEdit.id + ' successfully updated.');
                     this.closeModal();
 
                 }.bind(this))
                 .catch(function() {
-                    
+
                     this.notifyError('Product update error.');
                     this.closeModal();
-                    
+
                 }.bind(this));
             },
 
